# frontend/src/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SOC Analyzer",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/investigations": {
      "get": {
        "tags": [
          "investigations"
        ],
        "summary": "List Investigations",
        "description": "All cases, unresolved first, then worst first. `q` searches the case number, user, name and entities;\n`detector` keeps cases with at least one finding of that kind (Detection Rules \"View cases\").",
        "operationId": "list_investigations_api_investigations_get",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          },
          {
            "name": "severity",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "critical",
                    "high",
                    "medium",
                    "low"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Severity"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "open",
                    "investigating",
                    "resolved"
                  ],
                  "type": "string"
                },
                {
                  "const": "unresolved",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "detector",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Detector"
            }
          },
          {
            "name": "session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Session"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/InvestigationItem"
                  },
                  "title": "Response List Investigations Api Investigations Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/investigations/{case_id}": {
      "get": {
        "tags": [
          "investigations"
        ],
        "summary": "Get Investigation",
        "operationId": "get_investigation_api_investigations__case_id__get",
        "parameters": [
          {
            "name": "case_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Case Id"
            }
          },
          {
            "name": "session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Session"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvestigationDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "investigations"
        ],
        "summary": "Update Investigation",
        "operationId": "update_investigation_api_investigations__case_id__patch",
        "parameters": [
          {
            "name": "case_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Case Id"
            }
          },
          {
            "name": "session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Session"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CaseUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvestigationDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/investigations/{case_id}/notes": {
      "post": {
        "tags": [
          "investigations"
        ],
        "summary": "Add Note",
        "description": "Append-only: notes are never edited or deleted (add a new one to correct one).",
        "operationId": "add_note_api_investigations__case_id__notes_post",
        "parameters": [
          {
            "name": "case_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Case Id"
            }
          },
          {
            "name": "session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Session"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NoteIn"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvestigationDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CaseUpdate": {
        "properties": {
          "status": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "open",
                  "investigating",
                  "resolved"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Status"
          },
          "owner": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Owner"
          },
          "verdict": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "true_positive",
                  "false_positive",
                  "benign"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Verdict"
          },
          "expected_updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Expected Updated At"
          }
        },
        "type": "object",
        "required": [
          "expected_updated_at"
        ],
        "title": "CaseUpdate",
        "description": "Only the fields sent are changed; send `\"owner\": null` / `\"verdict\": null` to clear them."
      },
...
      "EntityChip": {
        "properties": {
          "type": {
            "type": "string",
            "enum": [
              "user",
              "device",
              "ip",
              "domain"
            ],
            "title": "Type"
          },
          "name": {
            "type": "string",
            "title": "Name"
          }
        },
        "type": "object",
        "required": [
          "type",
          "name"
        ],
        "title": "EntityChip"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "InvestigationItem": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "number": {
            "type": "string",
            "title": "Number"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "priority": {
            "type": "string",
            "enum": [
              "critical",
              "high",
              "medium",
              "low"
            ],
            "title": "Priority"
          },
          "risk": {
            "type": "integer",
            "title": "Risk"
          },
          "signals": {
            "type": "integer",
            "title": "Signals"
          },
          "alerts": {
            "type": "integer",
            "title": "Alerts"
          },
          "entities": {
            "items": {
              "$ref": "#/components/schemas/EntityChip"
            },
            "type": "array",
            "title": "Entities"
          },
          "status": {
            "type": "string",
            "enum": [
              "open",
              "investigating",
              "resolved"
            ],
            "title": "Status"
          },
          "owner": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Owner"
          },
          "verdict": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Verdict"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "start_ts": {
            "type": "string",
            "format": "date-time",
            "title": "Start Ts"
          },
          "end_ts": {
            "type": "string",
            "format": "date-time",
            "title": "End Ts"
          },
          "upload_id": {
            "type": "integer",
            "title": "Upload Id"
          }
        },
        "type": "object",
        "required": [
          "id",
          "number",
          "name",
          "title",
          "priority",
          "risk",
          "signals",
          "alerts",
          "entities",
          "status",
          "owner",
          "verdict",
          "updated_at",
          "start_ts",
          "end_ts",
          "upload_id"
        ],
        "title": "InvestigationItem"
      },
...
      "NoteIn": {
        "properties": {
          "text": {
            "type": "string",
            "maxLength": 5000,
            "minLength": 1,
            "title": "Text"
          }
        },
        "type": "object",
        "required": [
          "text"
        ],
        "title": "NoteIn"
      },
...
    }
  }
}